backend/requirements.txt
requests==2.26.0
django==3.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
psycopg2-binary
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-filter==21.1
python-dotenv==0.20.0
djoser==2.1.0
webcolors==1.11.1
drf-base64==2.0
